Load application .env file Allow to override publicPath by env variable (PUBLIC_PATH) Some utils moved to core
