+++ packages/core/src/utils/getRootDir.ts
