+++ packages/core/src/utils/resolvePath.ts
