+++ packages/webpack-config/src/index.ts
@@ -1,3 +1,2 @@
 export * from './WebpackConfig';
 export * from './WebpackConfigOptions';
-export * from './utils';
