--- packages/webpack-config/src/utils/index.ts
@@ -1,3 +1,0 @@
-export * from './getRootDir';
-export * from './resolvePath';
-export * from './resolveModule';
